% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{decompress_string}
\alias{decompress_string}
\title{Decode a PowerBI Compressed String}
\usage{
decompress_string(compressed_string)
}
\arguments{
\item{compressed_string}{The base64 string of the compressed data.}
}
\value{
Original string after decompression.
}
\description{
For more efficient storage, PowerBI will store strings in a compressed
binary format - compressed using the deflate algorithm and then represented as a base64 string.

This function takes the base64 string and decompresses it to return the original string.
}
\examples{
table_str <- "i45W8lXSUfJNzElVitWJVnIDctxSc2HcCCDXMS+/JCO1SKEktSgXLBgKFAzNy87LL88D8/2AfL/8EoXigtTkzLTM1BSl2FgA"
decompress_string(table_str) |> cat()
# [["M","Male"],["F","Female"],["X","Another term"],["U","Unknown"],["N","Not specified"]]
}
