% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerbi_dataflow.R
\name{download_dataflow_table}
\alias{download_dataflow_table}
\title{Download PowerBI Dataflow Table}
\usage{
download_dataflow_table(
  workspace_name,
  dataflow_name,
  table_name,
  access_token = NULL,
  destfile = NULL,
  verbose = TRUE
)
}
\arguments{
\item{workspace_name}{The PowerBI workspace name.}

\item{dataflow_name}{The name of the PowerBI Dataflow within the workspace.}

\item{table_name}{The name of the table within the PowerBI Dataflow to be
accessed.}

\item{access_token}{Token for authorising the connection to PowerBI. Valid options are:
\itemize{
\item \code{NULL}(default): Automatically generate via \code{qiverse.azure::get_az_tk('pbi_df')}
\item An \code{AzureAuth} object: Will be refreshed for the \verb{https://analysis.windows.net/powerbi/api} resource
\item A single string specifying the access token
}}

\item{destfile}{Optional output filepath to save raw dataflow CSV file to.
If \code{NULL} (default) the CSV is read into an R \code{data.frame}. If non-null,
the CSV file is downloaded to specified filepath and the function has no return.}

\item{verbose}{Whether to print status messages while the function is
running. Default is TRUE.}
}
\value{
A data.frame of the PowerBI table from the PowerBI dataflow if
\code{destfile = NULL}, otherwise none
}
\description{
This function provides the ability to download tables from within PowerBI
dataflow themselves, rather than requiring that they first be exposed via a
PowerBI dataset.
}
\examples{
 \dontrun{
# Create PowerBI Dataflow azure token
tk <- get_az_tk('pbi_df')

# Load AD dummy file from PowerBI dataflow
ad_dummy <- download_dataflow_table(
  workspace_name = "My Workspace Name",
  dataflow_name = "My Dataflow Name",
  table_name = "My Table Name",
  access_token = tk$credentials$access_token
)

# Download AD dummy file from PowerBI dataflow to local .csv file
download_dataflow_table(
  workspace_name = "My Workspace Name",
  dataflow_name = "My Dataflow Name",
  table_name = "My Table Name",
  access_token = tk$credentials$access_token,
  destfile = "my_table_name.csv"
)
}
}
