% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerbi_dataset_rest.R
\name{download_dataset_table_rest}
\alias{download_dataset_table_rest}
\title{This is the main function to be used. Specify the name of workspace, dataset,
and target table, and this function will query the API and return the table
as a data.frame}
\usage{
download_dataset_table_rest(
  workspace_name,
  dataset_name,
  table_name,
  access_token
)
}
\arguments{
\item{workspace_name}{The PowerBI workspace name.}

\item{dataset_name}{The name of the PowerBI Dataset within the workspace.}

\item{table_name}{The name of the table within the PowerBI Dataset to be
accessed.}

\item{access_token}{The token generated with the correct PowerBI Dataset
permissions. Use get_az_tk('pbi_ds') to create this token.}
}
\value{
A tibble of the PowerBI table from the PowerBI dataset.
}
\description{
The API endpoint is capped at returning a maximum of 1,000,000 values per
query, so we query the columns of the table in batches and then join them in
R
}
\examples{
 \dontrun{
# Create PowerBI Dataset azure token
tk <- get_az_tk('pbi_ds')

# Load AD dummy file from PowerBI dataflow
data_lbw <- download_dataset_table_rest(
  workspace_name = "My Workspace Name",
  dataset_name = "My Dataset (Semantic Model) Name",
  table_name = "My Table Name",
  access_token = tk$credentials$access_token)
}
}
