% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{execute_xmla_query}
\alias{execute_xmla_query}
\title{Execute a DAX query against a specified PowerBI Dataset using the XMLA API
endpoint.}
\usage{
execute_xmla_query(workspace, dataset, query, access_token = NULL)
}
\arguments{
\item{workspace}{Name of the workspace containing dataflow}

\item{dataset}{Name of the dataset to execute query against}

\item{query}{DAX query to execute}

\item{access_token}{Token for authorising the connection to PowerBI. Valid options are:
\itemize{
\item \code{NULL}(default): Automatically generate via \code{qiverse.azure::get_az_tk('pbi_df')}
\item An \code{AzureAuth} object: Will be refreshed for the \verb{https://analysis.windows.net/powerbi/api} resource
\item A single string specifying the access token
}}
}
\value{
DataFrame containing results of query
}
\description{
Execute a DAX query against a specified PowerBI Dataset using the XMLA API
endpoint.
}
