% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_pipelines.R
\name{get_pipeline_metadata}
\alias{get_pipeline_metadata}
\title{Request metadata for selected deployment pipeline name}
\usage{
get_pipeline_metadata(deployment_pipeline_name, access_token)
}
\arguments{
\item{deployment_pipeline_name}{Name of the deployment pipeline you want to
get metadata for}

\item{access_token}{The token generated with the correct PowerBI Dataset
permissions. Use get_az_tk('pbi_df') to create this token.}
}
\value{
DataFrame containing the names, GUIDs, and descriptions for the
selected deployment pipeline
}
\description{
Request metadata for selected deployment pipeline name
}
\examples{
 \dontrun{
# example code to get access token and list pipelines
token <- qiverse.azure::get_az_tk('pbi_df')
get_pipeline_metadata("my-pipeline", token$credentials$access_token)
}
}
