% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_pipelines.R
\name{list_pipelines}
\alias{list_pipelines}
\title{Request metadata for all deployment pipelines in Tenant}
\usage{
list_pipelines(access_token)
}
\arguments{
\item{access_token}{The token generated with the correct PowerBI Dataset
permissions. Use get_az_tk('pbi_df') to create this token.}
}
\value{
DataFrame containing the names, GUIDs, and descriptions for all
deployment pipelines in tenant you have access to.
}
\description{
Request metadata for all deployment pipelines in Tenant
}
\examples{
 \dontrun{
# example code to get access token and list pipelines
token <- qiverse.azure::get_az_tk('pbi_df')
list_pipelines(token$credentials$access_token)
}
}
