% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{list_workspaces}
\alias{list_workspaces}
\title{Request metadata for all Workspaces in Tenant}
\usage{
list_workspaces(access_token = NULL)
}
\arguments{
\item{access_token}{Token for authorising the connection to PowerBI. Valid options are:
\itemize{
\item \code{NULL}(default): Automatically generate via \code{qiverse.azure::get_az_tk('pbi_df')}
\item An \code{AzureAuth} object: Will be refreshed for the \verb{https://analysis.windows.net/powerbi/api} resource
\item A single string specifying the access token
}}
}
\value{
DataFrame containing the names, GUIDs, and capacity GUIDs for all workspaces in tenant.
}
\description{
Request metadata for all Workspaces in Tenant
}
