% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiverse.powerbi-package.R
\docType{package}
\name{qiverse.powerbi-package}
\alias{qiverse.powerbi-package}
\alias{qiverse.powerbi}
\title{The qiverse.powerbi package.}
\description{
A collection of R functions for working with PowerBI Dataflows and Datasets
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AUS-DOH-Safety-and-Quality/qiverse/qiverse.powerbi}
  \item Report bugs at \url{https://github.com/AUS-DOH-Safety-and-Quality/qiverse/issues}
}

}
\author{
\strong{Maintainer}: Healthcare Quality Intelligence Unit (Western Australia Health) \email{safetyandquality@health.wa.gov.au}

}
