% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerbi_dataflow.R
\name{refresh_dataflow}
\alias{refresh_dataflow}
\title{Refresh PowerBI Dataflow}
\usage{
refresh_dataflow(
  workspace_name,
  dataflow_name,
  access_token = NULL,
  verbose = TRUE
)
}
\arguments{
\item{workspace_name}{The PowerBI workspace name.}

\item{dataflow_name}{The name of the PowerBI Dataflow within the workspace.}

\item{access_token}{Token for authorising the connection to PowerBI. Valid options are:
\itemize{
\item \code{NULL}(default): Automatically generate via \code{qiverse.azure::get_az_tk('pbi_df')}
\item An \code{AzureAuth} object: Will be refreshed for the \verb{https://analysis.windows.net/powerbi/api} resource
\item A single string specifying the access token
}}

\item{verbose}{Whether to print status messages while the function is
running. Default is TRUE.}
}
\value{
A response object from the POST request.
}
\description{
This function provides the ability to refresh PowerBI dataflows, via an R
wrapper using the PowerBI REST API.
}
\examples{
 \dontrun{
# Create PowerBI Dataflow azure token
tk <- get_az_tk('pbi_df')

# Load AD dummy file from PowerBI dataflow
refresh_result <- refresh_dataflow(
  workspace_name = "My Workspace Name",
  dataflow_name = "My Dataflow Name",
  access_token = tk$credentials$access_token
)
}
}
