% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{table_string_to_json}
\alias{table_string_to_json}
\title{Decode a PowerBI Compressed Table String to JSON}
\usage{
table_string_to_json(table_str)
}
\arguments{
\item{table_str}{The base64 string representation of the table.}
}
\value{
A JSON string representing the table.
}
\description{
When entering data using the 'Enter Data' function in PowerBI, the
table is stored as JSON file, compressed, and the binary stored as a base64 string.

This function takes the base64 string and returns the JSON representation of the
table.
}
\examples{
table_str <- "i45W8lXSUfJNzElVitWJVnIDctxSc2HcCCDXMS+/JCO1SKEktSgXLBgKFAzNy87LL88D8/2AfL/8EoXigtTkzLTM1BSl2FgA"
table_string_to_json(table_str) |> cat()
# [["M","Male"],["F","Female"],["X","Another term"],["U","Unknown"],["N","Not specified"]]
}
