% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_dataset_table}
\alias{download_dataset_table}
\title{Download a specified dataset table into R using either the XMLA or REST API
endpoints. This is a convenience wrapper around the \code{execute_xmla_query()}
and \code{execute_rest_query()} functions.}
\usage{
download_dataset_table(
  workspace,
  dataset,
  table,
  method = "XMLA",
  access_token = NULL
)
}
\arguments{
\item{workspace}{Name of the workspace containing dataflow}

\item{dataset}{Name of the dataset containing table}

\item{table}{Name of the table to download}

\item{method}{The API to use for downloading the table. Valid values are "XMLA" (the default) and "REST"}

\item{access_token}{Token for authorising the connection to PowerBI. Valid options are:
\itemize{
\item \code{NULL}(default): Automatically generate via \code{qiverse.azure::get_az_tk('pbi_df')}
\item An \code{AzureAuth} object: Will be refreshed for the \verb{https://analysis.windows.net/powerbi/api} resource
\item A single string specifying the access token
}}
}
\value{
DataFrame containing downloaded table
}
\description{
Download a specified dataset table into R using either the XMLA or REST API
endpoints. This is a convenience wrapper around the \code{execute_xmla_query()}
and \code{execute_rest_query()} functions.
}
