% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{execute_arrow_query}
\alias{execute_arrow_query}
\title{Execute a DAX query against a specified PowerBI Dataset using the REST
Arrow API endpoint.}
\usage{
execute_arrow_query(workspace, dataset, query, access_token)
}
\arguments{
\item{workspace}{Name of the workspace containing dataflow}

\item{dataset}{Name of the dataset to execute query against}

\item{query}{DAX query to execute}

\item{access_token}{The token generated with the correct PowerBI Dataset
permissions. Use get_az_tk('pbi_ds') to create this token.}
}
\value{
DataFrame containing results of query
}
\description{
This endpoint has no restrictions on the size of the result and will
transfer data in a compressed format (Apache Arrow), and should be
preferred.
}
