% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{list_dataflows}
\alias{list_dataflows}
\title{Request metadata for all dataflows in specified workspace}
\usage{
list_dataflows(workspace, access_token)
}
\arguments{
\item{workspace}{Name of the workspace containing dataflows}

\item{access_token}{The token generated with the correct PowerBI Dataflow
permissions. Use get_az_tk('pbi_df') to create this token.}
}
\value{
DataFrame containing the names, GUIDs, and descriptions for all dataflows in workspace
}
\description{
Request metadata for all dataflows in specified workspace
}
