% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{list_workspaces}
\alias{list_workspaces}
\title{Request metadata for all Workspaces in Tenant}
\usage{
list_workspaces(access_token)
}
\arguments{
\item{access_token}{The token generated with the correct PowerBI Dataset
permissions. Use get_az_tk('pbi_ds') to create this token.}
}
\value{
DataFrame containing the names, GUIDs, and capacity GUIDs for all workspaces in tenant.
}
\description{
Request metadata for all Workspaces in Tenant
}
