% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerbi_dataflow.R
\name{update_dataflow_compute_engine}
\alias{update_dataflow_compute_engine}
\title{Update PowerBI Dataflow compute engine behaviour}
\usage{
update_dataflow_compute_engine(
  workspace_name,
  dataflow_name,
  compute_engine,
  access_token
)
}
\arguments{
\item{workspace_name}{The PowerBI workspace name.}

\item{dataflow_name}{The name of the PowerBI Dataflow within the workspace.}

\item{compute_engine}{The compute engine behaviour setting to be
applied. Must be one of "computeOptimized", "computeOn" or "computeDisabled".}

\item{access_token}{The token generated with the correct PowerBI Dataflow
permissions. Use get_az_tk('pbi_df') to create this token.}
}
\value{
A response object from the PATCH request.
}
\description{
This function provides the ability to change the compute engine behaviour for
a selected PowerBI dataflow, via an R wrapper using the PowerBI REST API.
}
\details{
You have to be the owner of the PowerBI dataflow to use this function,
otherwise there will be an 401 unauthorised error.
}
\examples{
 \dontrun{
# Create PowerBI Dataflow azure token
tk <- get_az_tk('pbi_df')

# Load AD dummy file from PowerBI dataflow
response <- update_dataflow_compute_engine(
  workspace_name = "My Workspace Name",
  dataflow_name = "My Dataflow Name",
  compute_engine = "computeOptimized",
  access_token = tk$credentials$access_token
)
}
}
