% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{update_dataset_storage_mode}
\alias{update_dataset_storage_mode}
\title{Update PowerBI Semantic Model (Dataset) storage mode}
\usage{
update_dataset_storage_mode(
  workspace_name,
  dataset_name,
  storage_mode,
  access_token
)
}
\arguments{
\item{workspace_name}{The PowerBI workspace name.}

\item{dataset_name}{The name of the PowerBI Dataflow within the workspace.}

\item{storage_mode}{The compute engine behaviour setting to be
applied. Must be one of "Abf" (small semantic model format) or "PremiumFiles"
(large semantic model format).}

\item{access_token}{The token generated with the correct PowerBI Dataflow
permissions. Use get_az_tk('pbi_df') to create this token.}
}
\value{
A response object from the PATCH request.
}
\description{
This function provides the ability to change the storage mode for a selected
PowerBI dataset, via an R wrapper using the PowerBI REST API.
}
\details{
You do not have to be the owner of the semantic model to use this function.
}
\examples{
 \dontrun{
# Create PowerBI Dataflow azure token
tk <- get_az_tk('pbi_df')

# Load AD dummy file from PowerBI dataflow
update_dataset_storage_mode <- refresh_dataflow(
  workspace_name = "My Workspace Name",
  dataset_name = "My Dataset Name",
  storage_mode = "PremiumFiles",
  access_token = tk$credentials$access_token
)
}
}
